Separate out some API components to alleviate drift between main API and demo API
